[Fix] Global volume slides were not limited properly when jumping around in a module. [Fix] Sample sync on jump could break if there was an EDx command in MOD/XM.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@4028 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ soundlib/Sndfile.cpp
@@ -872,7 +872,7 @@ bool CSoundFile::Create(FileReader file, ModLoadingFlags loadFlags)
 	if (!m_nDefaultSpeed) m_nDefaultSpeed = 6;
 	m_PlayState.m_nMusicSpeed = m_nDefaultSpeed;
 	m_PlayState.m_nMusicTempo = m_nDefaultTempo;
-	m_PlayState.m_nGlobalVolume = m_nDefaultGlobalVolume;
+	m_PlayState.m_nGlobalVolume = static_cast<int32>(m_nDefaultGlobalVolume);
 	m_PlayState.m_lHighResRampingGlobalVolume = m_PlayState.m_nGlobalVolume<<VOLUMERAMPPRECISION;
 	m_PlayState.m_nGlobalVolumeDestination = m_PlayState.m_nGlobalVolume;
 	m_PlayState.m_nSamplesToGlobalVolRampDest = 0;
